--- client/data/article/post.json
@@ -1,19 +1,0 @@
-[
-    {
-      "id": "1",
-      "author": "Marijn Haverbeke",
-      "published": "2018-12-04",
-      "image":"post.jpg",
-      "title": "Eloquent JavaScript, Third Edition",
-      "content": "JavaScript lies at the heart of almost every modern web application, from social apps like Twitter to browser-based game frameworks like Phaser and Babylon. Though simple for beginners to pick up and play with, JavaScript is a flexible, complex language that you can use to build full-scale applications.",
-      "tags": "tag1,tag2"
-    },
-    {
-        "id": "2",
-        "title": "Eloquent JavaScript, Third Edition",
-        "author": "Marijn Haverbeke",
-        "published": "2018-12-04",
-        "image":"post.jpg",
-        "content": "JavaScript lies at the heart of almost every modern web application, from social apps like Twitter to browser-based game frameworks like Phaser and Babylon. Though simple for beginners to pick up and play with, JavaScript is a flexible, complex language that you can use to build full-scale applications."
-      }
-  ]
