+++ client/data/post/hotpost.json
@@ -1,0 +1,1 @@
+[1,2,3]
